backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
borax==4.1.1
boto3==1.34.144
botocore==1.34.144
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.7
django-admin-geomap==1.6
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-echarts==0.6.0
django-filter==24.2
django-geojson==4.1.0
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-leaflet==0.30.1
django-map-widgets==0.5.1
django-modeltranslation==0.19.5
django-redis==5.4.0
django-storages==1.14.4
django-tinymce==4.1.0
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==22.0.0
htmlgenerator==1.2.30
idna==3.7
inflection==0.5.1
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.3.7
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
prettytable==3.10.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycparser==2.22
pyecharts==2.0.6
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
rabbitmq==0.2.0
redis==5.0.7
requests==2.32.3                        
s3transfer==0.10.2
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
daphne
databases
asyncpg
djangorestframework-simplejwt